Add recipe: Homemade Cinnamon Rolls
Homemade Cinammon Roll page containing an image.
Its linked from the homepage and the user may go back
using a link as well.
The steps part looks really crowded, maybe some CSS could fix this?

This will server as a template for the other recipes.
If this was a real website it could be autogenerated from markdown
to make adding new recipes more convenient.